pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yaml-config-support"
version = "0.1.0"
description = "Unterstützung für das Laden und Verwalten von YAML-Konfigurationsdateien"
authors = [
    { name="flow", email="groovehunter@posteo.de" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml",
    "click"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["yaml_config_support*"]
exclude = ["build*", "tests*"]
